+++ actioncable/test/subscription_adapter/redis_test.rb
@@ -19,12 +19,6 @@ class RedisAdapterTest < ActionCable::TestCase
   end
 end
 
-class RedisAdapterTest::Hiredis < RedisAdapterTest
-  def cable_config
-    super.merge(driver: "hiredis")
-  end
-end
-
 class RedisAdapterTest::AlternateConfiguration < RedisAdapterTest
   def cable_config
     alt_cable_config = super.dup
@@ -56,7 +50,7 @@ class RedisAdapterTest::ConnectorDefaultID < ActionCable::TestCase
   end
 
   test "sets connection id for connection" do
-    assert_called_with ::Redis, :new, [ expected_connection.stringify_keys ] do
+    assert_called_with ::Redis, :new, [ expected_connection.symbolize_keys ] do
       @adapter.send(:redis_connection)
     end
   end